+++ src/components/SortFilter.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SortFilter = () => {
+  return (
+    <div>SortFilter</div>
+  )
+}
+
+export default SortFilter;
