+++ src/state/reducers/commentsReducer.js
@@ -1,32 +1,12 @@
-import { bestreviewData } from '../../asset/bestreviewData';
-import { likeData } from '../../asset/likeData';
-import { recentData } from '../../asset/recentData';
+import { data } from '../../asset/data';
 
 const initialState = {
   loading: false,
   error: null,
   data: null,
   loaded: false,
+  length: 15
 };
-
-const GET_REVIEW_DETAIL = 'GET_REVIEW_DETAIL';
-const GET_REVIEW_DETAIL_SUCCESS = 'GET_REVIEW_DETAIL_SUCCESS';
-const GET_REVIEW_DETAIL_FAILURE = 'GET_REVIEW_DETAIL_FAILURE';
-
-export const getReviewDetail = (nickname) => ({
-  type: 'GET_REVIEW_DETAIL',
-  nickname,
-});
-
-export const getReviewDetailSuccess = ({ data }) => ({
-  type: 'GET_REVIEW_DETAIL_SUCCESS',
-  data,
-});
-
-export const getReviewDetaillFailure = (data) => ({
-  type: 'GET_REVIEW_DETAIL_FAILURE',
-  data,
-});
 
 export default function registerReducer(state = initialState, action) {
   switch (action.type) {
